split array.h into several headers
