+++ python/dgl/backend/pytorch/tensor.py
@@ -70,9 +70,10 @@ def context(input):
     return input.device
 
 def device_type(ctx):
-    return ctx.type
+    return th.device(ctx).type
 
 def device_id(ctx):
+    ctx = th.device(ctx)
     if ctx.index is None:
         return 0
     else:
